+++ src/YahooFinanceApi/Models/ChartResultIndicatorQuote.cs
@@ -2,11 +2,11 @@
 {
     public class ChartResultIndicatorQuote
     {
-        public double?[] Open { get; set; }
-        public double?[] Low { get; set; }
-        public double?[] High { get; set; }
+        public decimal?[] Open { get; set; }
+        public decimal?[] Low { get; set; }
+        public decimal?[] High { get; set; }
         
-        public double?[] Volume { get; set; }
-        public double?[] Close { get; set; }
+        public decimal?[] Volume { get; set; }
+        public decimal?[] Close { get; set; }
     }
 }
